refactor: move sendmail from init.py
